Back out changeset 7ece2f8e87f8: directly call certain Windows APIs.
The changeset in question was introduced specifically to make our build of the
library compatible with Windows XP, which lacks UpdateProcThreadAttribute(),
InitializeProcThreadAttributeList() and DeleteProcThreadAttributeList(). But
since we no longer care about Windows XP, and since this use of the
APR_DECLARE_LATE_DLL_FUNC() macro is spamming the build with opaque warnings,
undo the workaround.